15/04 7:54: sesions, router con middleware bastante guapo, contraseñas cifradas

+++ proyecto2ConTailwind/www/Core/Validator.php
@@ -3,7 +3,7 @@
 namespace Core;
 class Validator
 {
-    public static function checkString(string $text, int $minLenght = 1, int $maxLenght = INF)
+    public static function checkString(string $text, int $minLenght = 1, float $maxLenght = INF)
     {
         $text = trim($text);
